[refactor] create HazardList component
